ENG-4709 feat(portal,graphql): query consumer personal activity feed (#951)
- Reworks the personal activity feed (`/activity/personal`) to use
GraphQL with a few caveats. The data is scoped to the user and shows the
user's position on anything created (if there is one)
- Currently omits 'Deposited' and 'Redeemed' events due to some
questions about how to best filter these. I had Deposited events working
but omitting since this may be streamlined depending on how we decide to
answer the question around Redeemed events. We have another ticket for
this
- Pagination rework is also in another ticket, although the query
supports it and there is a temporary pagination component at the bottom
of the UI for testing purposes