ENG-4472 feat(portal,graphql): query consumer claim details (#957)
- Converts the Claim Details pages (index and id) to use GraphQL. Adds
the necessary queries to get the data as well as check if the Claim is
part of a List and also to pull the For/Against positions. Creates a new
version of the DetailInfoCard component to not break the Readonly route
build until we fully migrate and then we can remove
- Handling the positions was the biggest change from the previous
implementation. Previously, all positions came through in a single list
with a `for` and `against` direction as a property. This isn't the case
now, so we needed to combine the vault and counterVault positions
together. This works, but there are definitely performance improvements,
especially since we should likely be doing the tab filtering client side
instead of a refresh. There's room for us to continue to improve in the
future.
- Also, the search/sort/pagination will be wired up to the UI component
in a future PR, but the query supports it!
- @Vitalsine85 -- The staking modals will need to be updated as part of
the sweep you do. That should cover all the bases here!