ENG-4995 feat(1ui,graphql): update feature branch main (#970)
- Brings in crucial parts of the most recent PRs into `main` so that we
can continue working on migrating to the new schema/architecture on the
`feature/graphql-migration` branch
- This includes the logic in #969 -- we'll want to review that first so
any changes can also be reflected in here. Including that logic will
make working with the local API/schema more direct since we can
configure the URL a bit easier.
- I'm reasonably sure that I got everything. We'll have to deal with
this again when we merge this branch into `main` but now that the
GraphQL package is published #969 should be the last PR until we merge
this feature branch in. If we change approaches we can modify the
feature branch accordingly