ENG-4766 feat(graphql,portal): user profile data created (#971)
- This route update was challenging at the query and component level, so
a lot of changes in it. We'll be able to apply these to the other data
created routes though!
- Adds in/reworks positions, atoms, and triples queries to support the
data we need for each of the tabs
- Reworks the list/header components to work with the new data
structure, and adds logic for handling the positions for/against on each
triple
- Notes: We have a lot of values falling back to "Unknown" -- we're
aware of this and will be updated as the indexer is (this is using Dev
on Ponder). Also will update the pagination/sort/limits/etc. in another
ticket after fully changing to the new schema as we'll want this to be
locked in.