ENG-5006 fix(data-populator): atom existence bugs and db optimizations (#974)
There were:

Issues in how the database was remembering Atom IDs and their associated
URIs

This was consistently failing and falling back on batch EVM reads

Issues in how an existing atom was being checked when its info was
edited

The Type was not being provided, so edited atoms were checked as if
their full metadata was just  { https://schema.org/ }

Issues in how the database was being polled or written with large
batches of data (like a CSV full of atom data or their image src urls)

Batches were chunked into requests of 100, without accounting for the
length of each item - causing header overflow

This took a couple hours to resolve and track down as I was testing very
thoroughly and tracing all the internal steps.  But good thing we fixed
this.  This will:

Enhance stability and performance of Data Populator when checking if
atoms exist

Fix issues in Data Populator History Modal

Reduce excessive calls to Alchemy API (!!!)

This was detected by MK while trying to upload Berachain stuff. After he
published ~240 atoms they all showed up as green. He came back later and
loaded the same CSV to tag them, and only ~32 or so showed up as green
(published).

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)