ENG-5165 feat(repo): bootstrap new app setup (#984)
- Scaffolds a new `launchpad` in our apps/packages within monorepo
- Necessary routing only and no UI elements other than using the fonts
and theme
- Uses the newest versions of Remix and Vite along with opting in to
React Router v7 features
- Only pulls in the most needed packages and utilities for now as we
want to keep this fresh and only introduce as we need to
- Set up to use `1ui` and `graphql` packages -- GraphQL package
configured in the root to use a local API via the `configureClient`
configuration -- You'll need to run the API locally with this! The port
is set to `3000` in the `vite.config.ts` so there won't be collisions.
See my notes for a few more specific things regarding this.