ENG-5171 feat(launchpad,1ui): create activity feed and add to network route (#990)
## Affected Packages

Apps

- [ ] data populator
- [ ] portal
- [ ] template
- [x] launchpad

Packages

- [x] 1ui
- [ ] api
- [ ] graphql
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

Initially started building this in 1ui but then came to the conclusion
that it is tied too heavily into our logic for this to be a fit. Moved
into the launchpad and got it up on the Network route. Definitely needs
some more love and will have to work out some of the data with BE once
they're deployed.

## Screen Captures


![network](https://github.com/user-attachments/assets/eb2e3990-398d-41c2-be50-9a3540096e10)

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)