ENG-5293 feat(launchpad,1ui,graphql): bring protocol points query into graphql and basic scaffolding of points route (#1013)
## Affected Packages

Apps

- [ ] data populator
- [ ] portal
- [ ] template
- [x] launchpad

Packages

- [x] 1ui
- [ ] api
- [x] graphql
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

Brings the protocol points query into the graphql package. Basic
scaffolding of the points route with updated aggregated metrics
component. Query for relic points on points route.

Note: Tried setting up api package in launchpad so that we could query
for the other points, but realized that we'd have to set up privy for
auth. Will pick this back up after talking with the rest of the team.

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)

---------

Signed-off-by: Vital <rob@creative-element.net>