ENG-0000 feat(launchpad): patch knowledge graph (#1015)
## Affected Packages

Apps

- [ ] data populator
- [ ] portal
- [ ] template
- [x] launchpad

Packages

- [ ] 1ui
- [ ] api
- [ ] graphql
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

Should fix the vite issues when deployed to Render. Made some slight
changes to the knowledge graph layout to make connections easier to
read. Now when you hover a node or edge, it highlights all other
relevant nodes/edges. This caused hover cards not to work properly, but
will troubleshoot that later, as I think this change to highlighting is
more impactful.

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)