ENG-5433 ENG-5473 feat(launchpad): tcr table list with results, adds ctas to each page (#1039)
- Brings in `@tanstack/react-table` and connects to the game-1 wallet
list data. Adds a lot of sorting functionality, and ensures that the
values are based on the list. A lot of powerful scaffolding is added
with this, but we'll need to enhance the composability a bit to allow
for more adjustable columns. We'll do this as an enhancement as this
scaffolding gives us the power of React Table with our 1ui components.
- Updates the `AggregatedMetrics` component logic to calculate users and
tvl
- Adds CTAs to Play Again and View List
- Adds invalidation so the table automatically updates after users
complete the survey game