ENG-5594 fix(launchpad): fix validation errors in create step (#1066)
## Affected Packages

Apps

- [ ] data populator
- [ ] portal
- [ ] template
- [x] launchpad

Packages

- [ ] 1ui
- [ ] api
- [ ] graphql
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

## Overview

Validation errors weren't rendering on the create step, specifically if
a user had insufficient funds. Moved the check outside of the zod
schema. Should revisit to figure out ideal pattern, just wanted to get
this in to move QA forward.

## Screen Captures

If applicable, add screenshots or screen captures of your changes.

## Declaration

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)