ENG-5664 feat(launchpad,portal,graphql): bring signaling closer to portal state (#1079)
Apps

- [ ] data populator
- [x] portal
- [ ] template
- [x] launchpad

Packages

- [ ] 1ui
- [ ] api
- [x] graphql
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

Working off the previous PR, ports some of Portals staking components
into Launchpad. Updated the button components in the table rows, as well
as implemented the math to track the users ticks, or up-votes, and total
up-votes. This keeps things more in-line with the survey/question
experience. Users col now shows the number of up-votes on each entry, to
calculate this we do: `forTVL / minDeposit`. We can add downvotes when
we introduce staking against.

https://www.loom.com/share/f78f262c98344081a0e8edc5f5c9caee?sid=151eabf6-7c82-4b26-aecc-a7a0bb5a5911

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)