ENG-5695 fix(launchpad,graphql): temporary fix using signals feed for activity (#1090)
Apps

- [ ] data populator
- [ ] portal
- [ ] template
- [x] launchpad

Packages

- [ ] 1ui
- [ ] api
- [x] graphql
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

Created a GetSignalsQuery and am using signals feed on the activity now.
Note that this does not include creations, only deposits and withdraws,
but it works, so I see it as a temporary solution.

If applicable, add screenshots or screen captures of your changes.

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)