ENG-5665 feat(launchpad,1ui): implement downvoting (#1085)
Apps

- [ ] data populator
- [ ] portal
- [ ] template
- [x] launchpad

Packages

- [ ] 1ui
- [ ] api
- [ ] graphql
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

Introduces downvoting and further improves the signal/voting UX. Added
gradients to table and updated modal bg and border. Improved vault
details fetching. Removed unused components related to
survey/staking/signaling. Updated signal step in survey flow.

Note: still working through the vote math... 1 vote = minDeposit, but we
have to account for the fees when calculating it on the frontend. In the
case of transactions, 1 tick always = minDeposit, or if you're
redeeming, it will redeem all if parsedTick +1 > your shares. In the
case of showing someone their ticks post transaction and fees, we have
to do 0.94 * minDeposit to account for the 5% entry and 1% protocol
fees. Still seems to be some rounding issues though. Continuing to work
through that but we can test this now.

https://www.loom.com/share/655f4ec38b1a40baa0b56f3120c32188?sid=cfc75cc0-b176-4446-a321-02923a2bdadc

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)