ENG-5747 feat(launchpad,graphql): get relic/protocol points on launchpad rewards page (#1092)
Apps

- [ ] data populator
- [ ] portal
- [ ] template
- [x] launchpad

Packages

- [ ] 1ui
- [ ] api
- [ ] graphql
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

NOTE: Do not merge this in before the temporary activity fix, this is
based on that branch, sorry.

Pulls in relic points and protocol fees for protocol points on the
rewards page, in addition to general point aggregates. Eventually we
will use the points events table, but in the meantime we need a
temporary solution, and while this is fragmented it works. We just can't
get a leaderboard at the moment.

If applicable, add screenshots or screen captures of your changes.

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)