ENG-5683 feat(launchpad,1ui,graphql): update network page based on new designs (#1094)
Apps

- [ ] data populator
- [ ] portal
- [ ] template
- [x] launchpad

Packages

- [x] 1ui
- [ ] api
- [x] graphql
- [ ] protocol
- [ ] sdk

Tools

- [ ] tools

Note: This PR is based off of
https://github.com/0xIntuition/intuition-ts/pull/1093, do not merge this
in until that is merged in and this branch is rebased.

Reintroduces the Network route with updated aggregated metrics and
activity feed. Ported the activity feed from Portal into Launchpad.
Doesn't match the Koncepted designs exactly, but I don't think we need
to rebuild an entirely new feed, we just have to make a few more tweaks.

I am wondering if we want to modify the feed even further inside of
Launchpad to reflect upvotes, since we are using that terminology and
flow in the questions? This is a question for the larger group that I
will raise.

![image](https://github.com/user-attachments/assets/62de7c3b-c218-419d-9e6a-647b3a702fa7)

- [x] I hereby declare that I have abided by the rules and regulations
as outlined in the
[CONTRIBUTING.md](https://github.com/0xIntuition/intuition-ts/blob/main/CONTRIBUTING.md)